tambah2

+++ soal3.py
@@ -3,7 +3,7 @@ barang = int(input("masukan jumlah barang "))
 total = 0
 
 for i in range(barang):
-    harga = float(input("masukan harga "))
+    harga = int(input("masukan harga "))
     total += harga
 
 print("Total harga: ",total)
